Adjust alignment metadata for indentation persistence
This allows indentation to survive EditorState->html->EditorState transitions
with the updated version of htmlToDraftJs. The css is necessary to show the
indentation 'live', while the metadata gets translated into an inline style.

This is a companion to a change in htmlToDraftJs that translates an inline style
back into block metadata.

Reviewed by: melgar